requirements.txt
flwr==1.8.0
torch>=2.0.0
transformers==4.40.0
datasets==2.19.0
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
pandas==2.2.2
peft==0.10.0
evaluate==0.4.1
accelerate==0.29.3
tqdm==4.66.2
PyYAML==6.0.1
